+++ pubspec.yaml
@@ -1,10 +1,10 @@
 name: material_dialogs
 description: A Flutter package aims to help you create animated, simple, stylish Material Dialogs in your app.
-version: 1.1.3
+version: 1.1.4
 homepage: https://github.com/Ezaldeen99/material_dialogs
 
 environment:
-  sdk: '>=2.16.2 <3.0.0'
+  sdk: ">=2.16.2 <3.0.0"
 #  flutter: ">=1.17.0"
 
 dependencies:
@@ -12,7 +12,7 @@ dependencies:
     sdk: flutter
 
   # animations
-  lottie: ^1.3.0
+  lottie: ^2.1.0
 
 dev_dependencies:
   flutter_test:
